Add files via upload
This is my second task of the internship, I redesigned the Zara app I researched the Zara app and found the onboarding screen is not interactive and informative which leads to the user not understanding the app features. The app layout and interface are not visually appealing and confusing in navigation, leading to drop-off points when users use the app. I redesigned the app and made a user flow diagram for seamless navigation throughout the app.